feat: Add color visualization components for improved UI development
Implemented two new components, 'ColorsDescription' and 'TextColors', to enhance the color visualization process in our UI development. 'ColorsDescription' component dynamically displays color swatches alongside their corresponding names, improving the ease of identifying and selecting colors for design and development purposes.

The 'TextColors' component utilizes 'ColorsDescription' to showcase a curated list of colors used in the text elements of the application. This structured and visual representation of colors streamlines the design-to-development workflow, enabling quick reference and consistency in color usage across the application.

These components are integrated into our development environment using 'classnames' for dynamic styling and 'vibe-storybook-components' for an interactive display. This update significantly aids in maintaining color consistency and enhances the accessibility of our color palette for the development team.